+++ src/periodicorbit/PeriodicOrbitUtils.jl
@@ -3,7 +3,7 @@
 
 Update the continuation parameters according to a problem. This can be useful for branching from PD points where the linear solvers have to be updated, e.g. the number of unknowns is roughly doubled.
 """
-function _update_cont_params(contParams::ContinuationPar, pb::AbstractShootingProblem, orbitguess)
+function _update_cont_params(contParams::ContinuationPar, pb::AbstractPOShootingDiscretization, orbitguess)
     if contParams.newton_options.linsolver isa GMRESIterativeSolvers
         @reset contParams.newton_options.linsolver.N = length(orbitguess)
     end
@@ -17,7 +17,7 @@ function _update_cont_params(cont_params::ContinuationPar, coll::PeriodicOrbitOC
     return cont_params
 end
 
-@inline _update_cont_params(cont_params::ContinuationPar, pb::AbstractPOFDProblem, orbitguess) = cont_params
+@inline _update_cont_params(cont_params::ContinuationPar, pb::AbstractPOFiniteDifferencesDiscretization, orbitguess) = cont_params
 ####################################################################################################
 @inline user_passed_pofunction(rf::RecordForPeriodicOrbits) = user_passed_function(rf.user_record_from_solution)
 
@@ -30,22 +30,16 @@ function __user_record_solution_periodic_orbit(pbwrap, ::NoUserPassedFunction, i
     return (period = getperiod(pbwrap, getx(state), set(getparams(pbwrap), getlens(pbwrap), getp(state))),)
 end
 
-function __user_record_solution_periodic_orbit(pbwrap::AbstractWrapperFDProblem, ::UserPassedFunction, iter, state)
-    return pbwrap.recordFromSolution(getx(state), (prob = pbwrap.prob, p = getp(state)); iter, state)
-end
-
-function __user_record_solution_periodic_orbit(pbwrap::AbstractWrapperFDProblem, ::NoUserPassedFunction, iter::ContIterable{Tkind}, state) where {Tkind}
-    prob_po = pbwrap.prob
+function __user_record_solution_periodic_orbit(pbwrap::AbstractWrapperPOFiniteDifferencesProblem, ::NoUserPassedFunction, iter::ContIterable{Tkind}, state) where {Tkind}
+    prob_po = get_discretization(pbwrap)
     x = getx(state)
-    p = getp(state)
     period = getperiod(prob_po, x, nothing)
-    return (;period)
     sol = get_periodic_orbit(prob_po, x, nothing)
     _min, _max = @views extrema(sol[1, :])
     return (;max = _max, min = _min, amplitude = _max - _min, period)
 end
 
-function record_from_solution(iter::ContIterable{PeriodicOrbitCont, <: AbstractWrapperPOProblem},
+function record_from_solution(iter::ContIterable{PeriodicOrbitCont, <: AbstractWrapperPeriodicOrbitProblem},
                               state::AbstractContinuationState)
     probwrap = getprob(iter)
     __user_record_solution_periodic_orbit(probwrap, user_passed_pofunction(probwrap.recordFromSolution), iter, state)