+++ src/periodicorbit/StandardShooting.jl
@@ -56,7 +56,7 @@ or
 
 where we supply now two `ODEProblem`s. The first one `prob1`, is used to define the flow associated to `F` while the second one is a problem associated to the derivative of the flow. Hence, `prob2` must implement the following vector field ``\\tilde F(x,y,p) = (F(x,p), dF(x,p)\\cdot y)``.
 """
-@with_kw_noshow struct ShootingProblem{Tf <: AbstractFlow, Tjac <: AbstractJacobianType, Ts, Tsection, Tpar, Tlens} <: AbstractShootingProblem
+@with_kw_noshow struct ShootingProblem{Tf <: AbstractFlow, Tjac <: AbstractJacobianType, Ts, Tsection, Tpar, Tlens} <: AbstractPOShootingDiscretization
     "`ds`: vector of time differences for each shooting. Its length is written `M`. If `M == 1`, then the simple shooting is implemented and the multiple one otherwise."
     M::Int64 = 0                         # number of sections
     "`flow::Flow`: implements the flow of the Cauchy problem though the structure [`Flow`](@ref)."
@@ -123,11 +123,11 @@ end
 @inline get_time_slice(::ShootingProblem, x::AbstractMatrix, ii::Int) = @view x[:, ii]
 # orbit stored as [x1, x2, ...] where x1 is N-dimensional
 @inline get_time_slice(::ShootingProblem, x::AbstractVector, ii::Int) = x[ii]
-@inline get_time_slice(sh::ShootingProblem, x::AbstractVector{<:Real}, ii::Int) = get_time_slices(sh, x)[:, ii]
+@inline get_time_slice(sh::ShootingProblem, x::AbstractVector{ <: Real}, ii::Int) = get_time_slices(sh, x)[:, ii]
 @inline get_time_slice(::ShootingProblem, x::BorderedArray, ii::Int) = _getindex(x.u, ii)
 
 @inline get_time_slices(::ShootingProblem ,x::BorderedArray) = x.u
-@views function get_time_slices(sh::ShootingProblem, x::AbstractVector{<:Real})
+@views function get_time_slices(sh::ShootingProblem, x::AbstractVector{ <: Real})
     M = get_mesh_size(sh)
     N = div(length(x) - 1, M)
     return reshape(x[begin:end-1], N, M)